created component custom modal
