+++ src/components/PathwayCourse/index.js
@@ -16,18 +16,10 @@ import ExternalLink from "../common/ExternalLink";
 import LockIcon from "@mui/icons-material/Lock";
 import NoBatchEnroll from "../BatchClassComponents/NoBatchEnroll";
 import { CardContent } from "@mui/material";
-import OutlinedInput from "@mui/material/OutlinedInput";
-import InputLabel from "@mui/material/InputLabel";
-import MenuItem from "@mui/material/MenuItem";
-import FormControl from "@mui/material/FormControl";
-import ListItemText from "@mui/material/ListItemText";
-import Select from "@mui/material/Select";
-import Checkbox from "@mui/material/Checkbox";
-
 import { ReactComponent as CertificateIcon } from "./asset/certificate-grey.svg";
 import { ReactComponent as CertificateIconColored } from "./asset/certificate-color.svg";
 import Modal from "@mui/material/Modal";
-import CloseIcon from "@mui/icons-material/Close";
+import CustomModal from "./CustomModal";
 // import ReactPDF from "./ReactPDF.js";
 import {
   Container,
@@ -50,9 +42,6 @@ import axios from "axios";
 import { METHODS } from "../../services/api";
 import CustomSnackbar from "./customSnackbar";
 import MuiAlert from "@mui/material/Alert";
-import { StarRate } from "@material-ui/icons";
-
-import TextField from "@mui/material/TextField";
 import AmazonCodingProgrammer from "./AmazonCodingProgrammer";
 import { max } from "date-fns";
 
@@ -162,10 +151,7 @@ function PathwayCourse() {
 
   const [isFormModalOpen, setisFormModalOpen] = useState(false);
   const [certificate, setCertificate] = useState("");
-  let completedAll =
-    pathwayId == 8
-      ? completedPortion?.total >= 80
-      : completedPortion?.total === 100;
+  let completedAll = completedPortion?.total === 100;
   // let completedAll = true
   let [isFormFilled, setisFormFilled] = useState(false);
 
@@ -174,80 +160,8 @@ function PathwayCourse() {
   const displayCert = pathwayId == 1;
   // || pathwayId == 8;
 
-  let [teacherDetails, setTeacherDetails] = useState({
-    zone: "",
-    school_id: "",
-    school_name: "",
-    teacher_name: "",
-    teacher_id: "",
-    class_of_teacher: "",
-    email: "",
-  });
 
-  const [errors, setErrors] = useState({});
 
-  const validateForm = () => {
-    const newErrors = {};
-    if (!teacherDetails.zone) {
-      newErrors.zone = "Zone is required.";
-    }
-    if (!teacherDetails.school_id) {
-      newErrors.school_id = "School ID is required.";
-    } else if (teacherDetails.school_id.toString().length !== 7) {
-      newErrors.school_id = "School ID must be of 7 digits only";
-    }
-    if (!teacherDetails.school_name) {
-      newErrors.school_name = "School Name is required.";
-    }
-    if (!teacherDetails.teacher_name) {
-      newErrors.teacher_name = "Teacher Name is required.";
-    }
-    if (!teacherDetails.teacher_id) {
-      newErrors.teacher_id = "Teacher ID is required.";
-    } else if (teacherDetails.teacher_id.toString().length !== 8) {
-      newErrors.teacher_id = "Teacher ID must be of 8 digits only";
-    }
-    if (!teacherDetails.class_of_teacher) {
-      newErrors.class_of_teacher = "Class of Teacher is required.";
-    }
-    if (!teacherDetails.email) {
-      newErrors.email = "Email is required.";
-    } else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(teacherDetails.email)) {
-      newErrors.email = "Invalid email format.";
-    }
-    setErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  };
-
-  const handleSubmitteacherDetails = () => {
-    if (validateForm() === true) {
-      setLoader(true);
-      axios({
-        method: METHODS.POST,
-        url: `${process.env.REACT_APP_MERAKI_URL}/teacher/create`,
-        headers: {
-          accept: "application/json",
-          Authorization: user?.data?.token,
-        },
-        data: teacherDetails,
-      })
-        .then((res) => {
-          setLoader(false);
-          handleFormModalClose();
-          setisFormFilled(true);
-          setTeacherDetails({
-            zone: "",
-            school_id: "",
-            school_name: "",
-            teacher_name: "",
-            teacher_id: "",
-            class_of_teacher: "",
-            email: "",
-          });
-        })
-        .catch((err) => console.log(err));
-    }
-  };
 
   const modalStyle = {
     position: "absolute",
@@ -465,35 +379,7 @@ function PathwayCourse() {
   const handleFormModalClose = () => {
     setisFormModalOpen(false);
   };
-  const [teacherClass, setTeacherClass] = React.useState([]);
 
-  const handleChange = (event) => {
-    const {
-      target: { value },
-    } = event;
-    setTeacherClass(
-      typeof value === "string" ? value.split(",") : value
-    );
-  };
-  useEffect(()=>{
-    setTeacherDetails((prev)=>  ({
-      ...prev,
-      class_of_teacher: teacherClass.join(",")
-    }));
-  },[teacherClass])
-
-  const ITEM_HEIGHT = 48;
-  const ITEM_PADDING_TOP = 8;
-  const MenuProps = {
-    PaperProps: {
-      style: {
-        maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-        width: 250,
-      },
-    },
-  };
-  const classesOfTeacher = ["Class 1", "Class 2", "Class 3", "Class 4"];
-  const zoneArray = ["Central", "Civil Lines", "CTSP", "Karol Bagh", "Keshavpuram", "Narela", "Rohini", "Nazafgarh", "South", "Sharda.North", "Sharda.South", "West"]
 
 
   return (
@@ -509,175 +395,7 @@ function PathwayCourse() {
           </Alert>
         )}
       </Snackbar>
-      <Modal sx={{overflow:"scroll"}} open={isFormModalOpen} onClose={handleFormModalClose}>
-        <Box sx={style}>
-          <Box sx={{ display: "flex", justifyContent: "space-between" }}>
-            <Typography
-              id="modal-modal-title"
-              variant="h6"
-              component="h2"
-              sx={{
-                marginBottom: "2rem",
-              }}
-            >
-              Teacher Details
-            </Typography>
-            <CloseIcon
-              sx={{ cursor: "pointer" }}
-              onClick={handleFormModalClose}
-            />
-          </Box>
-
-          <Box
-            sx={{
-              display: "flex",
-              flexDirection: "column",
-              gap: "1rem",
-              marginBottom: "1rem",
-            }}
-          >
-            <FormControl fullWidth>
-              <InputLabel id="demo-simple-select-label">Select Zone</InputLabel>
-              <Select
-                label="Select Zone"
-                id="demo-simple-select"
-                value={teacherDetails.zone}
-                onChange={(e) => {
-                  setTeacherDetails((prev) => ({
-                    ...prev,
-                    zone: e.target.value,
-                  }));
-                }}
-              >
-                {zoneArray.map((name) => (
-                  <MenuItem key={name} value={name}>
-                    <ListItemText primary={name} />
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-            {errors.zone && (
-              <small style={{ color: "red" }}>{errors.zone}</small>
-            )}
-            <TextField
-              id="outlined-basic"
-              label="School Name"
-              variant="outlined"
-              value={teacherDetails.school_name}
-              onChange={(e) => {
-                e.persist();
-                setTeacherDetails((prev) => ({
-                  ...prev,
-                  school_name: e.target.value,
-                }));
-              }}
-            />
-            {errors.school_name && (
-              <small style={{ color: "red" }}>{errors.school_name}</small>
-            )}
-            <TextField
-              id="outlined-basic"
-              label="School Id"
-              type="number"
-              variant="outlined"
-              value={teacherDetails.school_id}
-              onChange={(e) => {
-                e.persist();
-                setTeacherDetails((prev) => ({
-                  ...prev,
-                  school_id: parseInt(e.target.value),
-                }));
-              }}
-            />
-            {errors.school_id && (
-              <small style={{ color: "red" }}>{errors.school_id}</small>
-            )}
-            <TextField
-              id="outlined-basic"
-              label="Teacher Name"
-              variant="outlined"
-              value={teacherDetails.teacher_name}
-              onChange={(e) => {
-                e.persist();
-                setTeacherDetails((prev) => ({
-                  ...prev,
-                  teacher_name: e.target.value,
-                }));
-              }}
-            />
-            {errors.teacher_name && (
-              <small style={{ color: "red" }}>{errors.teacher_name}</small>
-            )}
-            <TextField
-              id="outlined-basic"
-              label="Email "
-              variant="outlined"
-              value={teacherDetails.email}
-              onChange={(e) => {
-                e.persist();
-                setTeacherDetails((prev) => ({
-                  ...prev,
-                  email: e.target.value,
-                }));
-              }}
-            />
-            {errors.email && (
-              <small style={{ color: "red" }}>{errors.email}</small>
-            )}
-            <TextField
-              id="outlined-basic"
-              label="Teacher ID"
-              type="number"
-              variant="outlined"
-              value={teacherDetails.teacher_id}
-              onChange={(e) => {
-                e.persist();
-                setTeacherDetails((prev) => ({
-                  ...prev,
-                  teacher_id: parseInt(e.target.value),
-                }));
-              }}
-            />
-            {errors.teacher_id && (
-              <small style={{ color: "red" }}>{errors.teacher_id}</small>
-            )}
-            <FormControl sx={{ width:max }}>
-              <InputLabel id="demo-multiple-checkbox-label">
-                Select Class
-              </InputLabel>
-              <Select
-                labelId="demo-multiple-checkbox-label"
-                id="demo-multiple-checkbox"
-                multiple
-                value={teacherClass}
-                onChange={handleChange}
-                input={<OutlinedInput label="Select Class" />}
-                renderValue={(selected) => selected.join(", ")}
-                MenuProps={MenuProps}
-              >
-                {classesOfTeacher.map((name) => (
-                  <MenuItem key={name} value={name}>
-                    <Checkbox checked={teacherClass.indexOf(name) > -1} />
-                    <ListItemText primary={name} />
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-            {errors.class_of_teacher && (
-              <small style={{ color: "red" }}>{errors.class_of_teacher}</small>
-            )}
-          </Box>
-          <Button
-            variant="contained"
-            sx={{
-              marginLeft: "60%",
-            }}
-            onClick={handleSubmitteacherDetails}
-          >
-            Share Details
-          </Button>
-        </Box>
-      </Modal>
+      <CustomModal isFormModalOpen={isFormModalOpen} setisFormFilled={setisFormFilled} setisFormModalOpen={setisFormModalOpen}  user={user}/>
       {pathwayId === "7" ? (
         <AmazonCodingProgrammer pathwayId={pathwayId} />
       ) : (
